[ refactor ] 폴더 경로 수정 및 파일 명 변경
