--- components/button/iconButton.tsx
@@ -1,19 +1,0 @@
-import Image from 'next/image';
-
-interface IconButtonProps {
-  src: string;
-  alt: string;
-  onClick?: () => void;
-}
-
-export default function IconButton(props: IconButtonProps) {
-  const { src, alt, onClick } = props;
-
-  const handleClick = () => {
-    if (onClick) {
-      onClick();
-    }
-  };
-
-  return <Image onClick={handleClick} src={src} alt={alt} style={{ cursor: 'pointer' }} />;
-}
