+++ components/common/modal/termsModal.tsx
@@ -1,7 +1,7 @@
 import theme from '@/styles/theme';
 import styled from 'styled-components';
 import Image from 'next/image';
-import IconButton from '@/components/button/iconButton';
+import IconButton from '@/components/common/button/iconButton';
 import { CloseWhiteIc, DetailCloseIc, DetailOpenIc } from '@/public/assets/icons';
 import { useEffect, useState } from 'react';
