+++ pages/cakes/[id].tsx
@@ -7,9 +7,9 @@ import InputContainer from '@/components/common/input/inputContainer';
 import InputTitle from '@/components/common/input/inputTitle';
 import TextareaBox from '@/components/common/input/textareaBox';
 import { CAKE_LIST } from '@/constant/cakeList';
-import { CakeListType } from '@/types/cakeListType';
+import { CakeListType } from '@/types/cakes/cakeListType';
 import Image from 'next/image';
-import ButtonBox from '@/components/button/buttonBox';
+import ButtonBox from '@/components/common/button/buttonBox';
 import { LIMIT_TEXT } from '@/constant/limitText';
 import { useEffect, useState } from 'react';
 import CakesHeader from '@/components/cakes/cakesHeader';
