+++ pages/wishes/[id].tsx
@@ -7,7 +7,7 @@ import Header from '@/components/common/header';
 import IconButton from '@/components/common/button/iconButton';
 import { useEffect, useState } from 'react';
 import GuideModal from '@/components/common/modal/GuideModal';
-import ButtonBox from '@/components/button/buttonBox';
+import ButtonBox from '@/components/common/button/buttonBox';
 import { useRouter } from 'next/router';
 import { useQuery } from 'react-query';
 import { getWishesData } from '@/api/cakes/getWishesData';
